Do CPR analysis even for sum types
Based on a draft patch by Max Bolingbroke (Trac #5075), this
commit makes demand analysis do CPR on sum types.  The idea is
that if a function returns only constructor of a sum type (eg it
returns only Just, never Nothing, from a Maybe type), we can
do CPR analysis on it.  This may be somewhat unusual, but it does
happen (Trac #5075).

The fundamental change is that the DmdResType records the data
constructor

   data DmdResultP con
     = TopRes
     | RetCPR con    -- The 'con' field is new
     | BotRes

The same type DmdResultP is used in Core (with con = DataCon)
and in IfaceSyn (with con = IfExtName) to avoid duplication.

Beyond this, the patch adds a bit more info to the DmdResType, and
that's a direction I want to go in for doing *nested* CPR (Trac #1600).

Finally, I'm pretty sure we can cure the current problem with CPR for
data families, by going a tiny bit further.

In doing this patch I also made a useful simplification:

  * Remove the Maybe from the strictnessInfo and demandInfo
    fields of IdInfo

The Maybe was there to record the fact that this was the "first time"
through in the demand analyser; ie the existing strictness/demand
should be ignored.  But that is now handled much more tidily (and more
locally) by the ae_virgin field of the AnalEnv, in DmdAnal.  A raft
of small simplifications follow.