Make the canonicaliser return StopOrContine
In the very common case that the work item is already
canonical we don't want to stuff it back in the work list
and start the pipeline again; it's a waste.  Actually
the change worked out rather nicely.

I also changed PredTree to be a *single level* of PredType
classification rather than recursive.  Less to-and-fro stuff
now.