Change the compiler's internal representation of vectors.
Previously, a vector's representation was indexed by a representation meant to
reflect the type of elements in the vector. However, we need to be explicit
about the size of a vector's elements (particularly for Ints and Words), which
the existing representation didn't allow. The representation of a vector's
elements is now separate from the representation of "normal" values.