Handle 4-byte aligned stack on Win32 when generating SSE instructions.
Win32 only guarantees that the stack is 4-byte aligned, so rewrite all movaps
instructions to movups in the mangler on that platform. We already generate
movups for explicit loads and stores in the LLVM back-end because they are
marked as potentially unaligned. However, LLVM generates movaps for registers
spills, and I don't see any way to fix that except by rewriting the assembly
output in the mangler.