Add forward reasoning: derive new givens from existing ones.
This seems to work, although it'd be nice to control the firing
of rules a bit more as at the moment we generate a lot of useless
facts.  For example, given 2 + 3 ~ a, we get all of these:

3 + 2 ~ a
a ~ 5
5 ~ a

Some of this can be fixed by restricting the instantiations of
"funRules", which always produce symmetric results.