Improve representation of AxiomRules.
This commit also fixes the representation of axiom rules in interfaces,
and the checking of axiom rules in the core lint.

Now we have two shape of axiom rule:
  * The first is the usual one: "forall as. ps => q"
  * The new one provides a way to specify axiom-schemas, which are
     useful when defining functions on type literals.

For example, to define addition on numeric literals we use a
rule schema, AddDef, which behaves like this:

  AddDef@(1,2) :: 1 + 2 ~ 3
  AddDef@(2,3) :: 2 + 3 ~ 5
  etc.

The schema has 2 type parameters (which have to be type literals)
and asserts that the definition of type function (+) behaves like addition.