Prefer Wanteds over Derived when solving family equations:
If we have a choice of a wanted and derived equation, we prefer
the wanted one.  To see why, consider the following example:

        [D] b + a ~ c
        [W] a + b ~ c

If we use the derived one, then `c` gets defined to `b + a` and we
are left with an unsolved wanted constraint because now `c` has
a binding.  However, if we use the wanted first, then we are
left with an "unsolved" derived constraint, which is OK because
derived constraints don't correspond to goals that need to be solved.
(Indeed, it is likely that the derived constraint was generated
by the wanted, to enable reactions where the arguments
to (+) were swapped)