Add custom improvement rule for `FromNat1`.
When we see `FromNat1 x ~ y` and we can prove that `1 <= y` we generate
a derived contraint `x = Succ a` from a new flexible variable `a`.

This follows from the injectivity of the function.

Note that this is similar to what happens with fun.deps. in GHC:
it should work fine when we are doing type inference but is likely
to fail with given constraints because "derived" facts do not interact
with givens.

It is interesting that we could add a rule to do this properly
(i.e., with a proof) but the rule would need an existential (i.e., would
introduce a skolem variable) and e don't have support for this at present:

forall a b. exists c. (FromNat1 a ~ b, 1 <= b) => a ~ Succ c