Fixes to restore proper operation of widening function.
We need to be careful when we use BRules in interaction:
unlike axiom rules, using them does not always result in
useful information---in fact often it results in trivial facts.

Furthermore, we have to be careful not to disable an interaction
with an existing assumption (i.e., if we apply a rule to bRule,
we should also keep a copy around in case the same slot might
be filled by an assumption).

At present, we only interact the BRules after we've tried all assumptions:
this reduces the number of useless facts we construct, but is sufficient
for the main use case of interacting with BRules, namely improvements
of the form: a + 0 ~ b => a ~ b.